Changes to registration page and side menu.
The side menu now only displays the text "Sign in with any of the following" when the user is not signed in. Otherwise, the header to the side menu says "Main Menu:". The URL from the registration page now only changes if all the information entered into the fields was valid.